Refactor blogs
