+++ pcweb/components_webpage/sidebar/sidebar.py
@@ -8,7 +8,7 @@ from pcweb import constants
 from pcweb.pages.docs import getting_started, hosting
 from pcweb.pages.faq import faq
 from pcweb.pages.docs.library import library
-from pcweb.pages.blog import blg
+from pcweb.pages.blog import blogs
 from pcweb.pages.changelog import changelog
 from pcweb.pages.docs.gallery import gallery
 
@@ -22,7 +22,7 @@ def sidebar(url=None, width: str = "100%") -> rx.Component:
     }
 
     return rx.vstack(
-        rx.link("Blog", href=blg.path, style=section_style),
+        rx.link("Blog", href=blogs.path, style=section_style),
         rx.link("Gallery", href=gallery.path, style=section_style),
         rx.link("Docs", href=getting_started.introduction.path, style=section_style),
         rx.link("Component Library", href=library.path, style=section_style),
