+++ pcweb/pages/blog/__init__.py
@@ -1,0 +1,1 @@
+from .blog import *
